pyproject.toml
[tool.poetry]
name        = "yoloflow"
version     = "0.1.2"
description = "TouchDesigner-oriented YOLO workflow tool"
authors     = ["Philliddellphia"]
license     = "MIT"
readme      = "README.md"
packages    = [{ include = "yoloflow" }]

[tool.poetry.dependencies]
python       = "^3.11"

# CUDA-enabled PyTorch (11.8 tool-chain)
torch        = { version = "2.7.0+cu118",  source = "pytorch-cu118" }
torchvision  = { version = "0.22.0+cu118", source = "pytorch-cu118" }

# GUI
PyQt5        = "5.15.11"

# Core libs
opencv-python = "4.11.0.86"
pillow        = "11.2.1"
pyyaml        = "6.0.2"
cyndilib      = "0.0.5"
numpy         = "2.2.5"
python-osc    = "1.9.3"
setuptools    = "80.3.0"
timm          = "1.0.15"
tqdm          = "4.67.1"
ultralytics   = "8.3.126"

[[tool.poetry.source]]
name     = "pytorch-cu118"
url      = "https://download.pytorch.org/whl/cu118"
priority = "explicit"           # only the packages referenced above come from here

[tool.poetry.scripts]
yoloflow = "yoloflow.main:main"

[build-system]
requires = ["poetry-core>=1.7.0"]   # 1.7+ avoids old “default”/source bugs
build-backend = "poetry.core.masonry.api"
